+++ emapp/src/wasm_init.rs
@@ -101,7 +101,7 @@ impl WasmClient {
                     window_device_pixel_ratio: 1.0f32,
                     viewport_device_pixel_ratio: 1.0f32,
                     window_size: [1, 1],
-                    viewport_size: [size.width as u16, size.height as u16],
+                    viewport_size: [size.width, size.height],
                 },
             );
